refactor: move ptx custom handlers in separate file Add ptx custom handlers to WebPlatformPlayer for discover section
