+++ apps/ledger-live-desktop/src/renderer/components/WebPlatformPlayer/index.tsx
@@ -7,6 +7,7 @@ import { TopBar, TopBarConfig } from "./TopBar";
 import Box from "../Box";
 import { WebviewAPI, WebviewProps, WebviewState } from "../Web3AppWebview/types";
 import { initialWebviewState } from "../Web3AppWebview/helpers";
+import { usePTXCustomHandlers } from "../WebPTXPlayer/CustomHandlers";
 
 export const Container = styled.div`
   display: flex;
@@ -34,11 +35,14 @@ export default function WebPlatformPlayer({ manifest, inputs, onClose, config, .
   const webviewAPIRef = useRef<WebviewAPI>(null);
   const [webviewState, setWebviewState] = useState<WebviewState>(initialWebviewState);
 
+  const customPTXHandlers = usePTXCustomHandlers(manifest);
+
   const customHandlers = useMemo<WalletAPICustomHandlers>(() => {
     return {
       ...loggerHandlers,
+      ...customPTXHandlers,
     };
-  }, []);
+  }, [customPTXHandlers]);
 
   const onStateChange: WebviewProps["onStateChange"] = state => {
     setWebviewState(state);
