+++ apps/ledger-live-mobile/src/components/WebPlatformPlayer/index.tsx
@@ -1,10 +1,11 @@
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { StyleSheet, SafeAreaView, BackHandler, Platform } from "react-native";
 
 import { useNavigation } from "@react-navigation/native";
 import { Flex } from "@ledgerhq/native-ui";
 import { safeGetRefValue } from "@ledgerhq/live-common/wallet-api/react";
+import { handlers as loggerHandlers } from "@ledgerhq/live-common/wallet-api/CustomLogger/server";
 import { WebviewAPI, WebviewState } from "../Web3AppWebview/types";
 
 import { Web3AppWebview } from "../Web3AppWebview";
@@ -15,6 +16,8 @@ import { BaseNavigatorStackParamList } from "../RootNavigator/types/BaseNavigato
 import HeaderTitle from "../HeaderTitle";
 import { initialWebviewState } from "../Web3AppWebview/helpers";
 import { InfoPanel } from "./InfoPanel";
+import { usePTXCustomHandlers } from "../WebPTXPlayer/CustomHandlers";
+import { WalletAPICustomHandlers } from "@ledgerhq/live-common/wallet-api/types";
 
 type Props = {
   manifest: LiveAppManifest;
@@ -72,6 +75,15 @@ const WebPlatformPlayer = ({ manifest, inputs }: Props) => {
     });
   }, [manifest, navigation, webviewState]);
 
+  const customPTXHandlers = usePTXCustomHandlers(manifest);
+
+  const customHandlers = useMemo<WalletAPICustomHandlers>(() => {
+    return {
+      ...loggerHandlers,
+      ...customPTXHandlers,
+    };
+  }, [customPTXHandlers]);
+
   return (
     <SafeAreaView style={[styles.root]}>
       <Web3AppWebview
@@ -79,6 +91,7 @@ const WebPlatformPlayer = ({ manifest, inputs }: Props) => {
         manifest={manifest}
         inputs={inputs}
         onStateChange={setWebviewState}
+        customHandlers={customHandlers}
       />
       <BottomBar manifest={manifest} webviewAPIRef={webviewAPIRef} webviewState={webviewState} />
       <InfoPanel
